refactor(ast): extract nodeToString
